fix regression with boost fusion 1.58
The fold operation in boost fusion 1.58 seems to be more strict when it
comes to the types of its arguments. Before, fold would properly
propagate the types, when supplied with a list of size_t and an int as
initial state. Also it was a bit less strict with the types of the
arguments of its function-argument.

Now we make sure to only feed size_t values to fold and we changed the
first argument of offsetHelper and globalOffsetHelper to const size_t&
instead of size_t.

+++ dune/dpg/innerproduct.hh
@@ -120,7 +120,7 @@ namespace Dune {
 
       /* set up local offsets */
       localTotalTestSize =
-          fold(zip(localTestSpaceOffsets, tlv), 0, offsetHelper());
+          fold(zip(localTestSpaceOffsets, tlv), (size_t)0, offsetHelper());
     }
 
     /**
@@ -173,7 +173,7 @@ getOccupationPattern(MatrixIndexSet& nb) const
   /* set up global offsets */
   size_t globalTestSpaceOffsets[std::tuple_size<TestSpaces>::value];
   fold(zip(globalTestSpaceOffsets, testBasisIndexSet),
-       0, globalOffsetHelper());
+       (size_t)0, globalOffsetHelper());
 
   auto testLocalView = as_vector(transform(testSpaces,
                                            localViewFromFEBasis()));
